Update README.md
INTRODUCTION 
This project will be conducting text mining and sentiment analysis on a dataset containing thousands of reviews from customers about their experiences in various hotels and restaurants in Thailand. 
The main objective of this part is to select 30 hotels/restaurants and perform show how sentiment analysis and text mining can be performed using python and classify sentiments based on the reviews as negative, positive and neutral.  
According to linguamaics.com  “Text mining is an artificial intelligence technology that uses natural language processing (NLP) to transform unstructured text into a structured format to identify meaningful patterns and insights.” While “Sentiment analysis is a technique that detects the underlying sentiment in a piece of text and classifies them as either positive, negative or neutral.” 
 
THE DATASET
The dataset that will be used for this part is a selection of reviews of 30 hotels/Restaurants from two of the major districts in the Phuket province. This data selection was done in Microsoft Excel before importing to python. (The data set is an excel file which has been submitted alongside this assessment. 
The Original tourism_accomodation_reviews.csv dataset consists of customer reviews of hotels and restaurants in 25 different locations in the province of Phuket, Thailand. 
We chose 15 hotels/Restaurants from Choeng thale in Thalang, a district in the Northern part of Phuket and another 15 hotels/Restaurants in Cape Panwa which is situated in Mueang Phuket District in the southern part of Phuket and also the capital of the province. 
Phuket is one of the southern provinces of Thailand. It is made up of the Phuket Island which is the largest island in the country and 32 other smaller islands. 
The climate of Phuket features a tropical monsoon climate and the city has an average annual high temperature of 32 °C (90 °F) and an annual low of 25 °C (77 °F). 
Tin mining was the major source of income for the in the 16th century but the economy now rests on two major pillars which is the rubber tree plantations and Tourism. 
Although these two locations are at two opposite ends of the Phuket Province, they are both great for Sightseeing tours, good resorts, spas, local eats and drinks, water sports and after-dark activities.